% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.mvQCAbt.R
\name{summary.mvQCAbt}
\alias{summary.mvQCAbt}
\title{Summarize Binomial Tests for mvQCA Data}
\usage{
\method{summary}{mvQCAbt}(object, ...)
}
\arguments{
\item{object}{Object returned by \code{\link{mvQCAbinTest}}.}

\item{...}{Additional parameters to pass on.}
}
\value{
Matrix of values for counterexamples and consistency.
}
\description{
Displays number of confirming cases and raw and adjusted p-scores
following Binomial test of mvQCA data.
}
\examples{
test <- mvQCAbinTest(freq.y=0.7, configs=list(aB=5, bCD=3, Ce=2),
   total.configs=20)
summary(test)
}
\keyword{binomial}
\keyword{mvQCA}
\keyword{test}

